Move enhanced agenda content out of conflicted page file
