+++ src/pages/components/AgendaContent.tsx
@@ -1,13 +1,8 @@
-import { useEffect, useMemo, useState } from "react";
-import { View } from "react-big-calendar";
-import { addDays, endOfDay, endOfMonth, endOfWeek, format, startOfDay, startOfMonth, startOfWeek, subDays } from "date-fns";
-import { ptBR } from "date-fns/locale";
+import { useMemo, useState } from "react";
 import { useAuth } from "@/hooks/useAuth";
 import { supabase } from "@/integrations/supabase/client";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
 import { useToast } from "@/hooks/use-toast";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
@@ -20,62 +15,17 @@ import { AppointmentDetailsDialog } from "@/components/agenda/AppointmentDetails
 import ImportAppointmentsDialog from "@/components/agenda/ImportAppointmentsDialog";
 import { STATUS_LABELS, STATUS_VARIANTS, STATUS_OPTIONS, normalizeStatus } from "@/lib/appointmentStatus";
 
-type PeriodMode = "day" | "week" | "month" | "custom";
-type Professional = { id: string; name: string };
-
-const ALL_PROFESSIONALS = "all";
-const FILTER_STORAGE_KEY = "agenda.professionalFilter";
-const APPOINTMENT_ID_BATCH_SIZE = 500;
-
-const getWeekOptions = () => ({ locale: ptBR, weekStartsOn: 0 as const });
-
-function getRangeForPeriod(periodMode: PeriodMode, date: Date) {
-  if (periodMode === "day") {
-    return { start: startOfDay(date), end: endOfDay(date) };
-  }
-
-  if (periodMode === "month") {
-    return { start: startOfMonth(date), end: endOfMonth(date) };
-  }
-
-  return {
-    start: startOfWeek(date, getWeekOptions()),
-    end: endOfWeek(date, getWeekOptions()),
-  };
-}
-
-function toDateInputValue(date: Date) {
-  return format(date, "yyyy-MM-dd");
-}
-
-function fromDateInputValue(value: string, fallback: Date) {
-  if (!value) return fallback;
-  const [year, month, day] = value.split("-").map(Number);
-  if (!year || !month || !day) return fallback;
-  return new Date(year, month - 1, day);
-}
-
-function formatRangeLabel(start: Date, end: Date) {
-  return `${format(start, "dd/MM/yyyy")} até ${format(end, "dd/MM/yyyy")}`;
-}
-
 export default function AgendaContent() {
-  const { profile, establishmentRole, professionalId } = useAuth();
+  const { profile } = useAuth();
   const { toast } = useToast();
   const qc = useQueryClient();
   const establishmentId = profile?.id as string | undefined;
-  const isEmployee = establishmentRole === "employee";
 
   const [viewMode, setViewMode] = useState<"calendar" | "list">("calendar");
-  const [calendarView, setCalendarView] = useState<View>("week");
-  const [calendarDate, setCalendarDate] = useState<Date>(new Date());
-  const [periodMode, setPeriodMode] = useState<PeriodMode>("week");
-  const [range, setRange] = useState<{ start: Date; end: Date }>(() => getRangeForPeriod("week", new Date()));
-  const [customStart, setCustomStart] = useState(() => toDateInputValue(range.start));
-  const [customEnd, setCustomEnd] = useState(() => toDateInputValue(range.end));
-  const [selectedProfessionalId, setSelectedProfessionalId] = useState<string>(() => {
-    if (typeof window === "undefined") return ALL_PROFESSIONALS;
-    return window.localStorage.getItem(FILTER_STORAGE_KEY) || ALL_PROFESSIONALS;
+  const [range, setRange] = useState<{ start: Date; end: Date }>(() => {
+    const s = new Date(); s.setDate(s.getDate() - s.getDay()); s.setHours(0, 0, 0, 0);
+    const e = new Date(s); e.setDate(e.getDate() + 6); e.setHours(23, 59, 59, 999);
+    return { start: s, end: e };
   });
   const [formOpen, setFormOpen] = useState(false);
   const [detailsOpen, setDetailsOpen] = useState(false);
@@ -83,163 +33,29 @@ export default function AgendaContent() {
   const [selectedAppt, setSelectedAppt] = useState<any | null>(null);
   const [initialSlot, setInitialSlot] = useState<Date | null>(null);
 
-  const effectiveProfessionalId = isEmployee ? professionalId : selectedProfessionalId !== ALL_PROFESSIONALS ? selectedProfessionalId : null;
-
-  useEffect(() => {
-    if (!isEmployee || !professionalId) return;
-    setSelectedProfessionalId(professionalId);
-  }, [isEmployee, professionalId]);
-
-  useEffect(() => {
-    if (isEmployee || typeof window === "undefined") return;
-    window.localStorage.setItem(FILTER_STORAGE_KEY, selectedProfessionalId);
-  }, [isEmployee, selectedProfessionalId]);
-
-  useEffect(() => {
-    if (periodMode !== "custom") {
-      const nextRange = getRangeForPeriod(periodMode, calendarDate);
-      setRange(nextRange);
-      setCustomStart(toDateInputValue(nextRange.start));
-      setCustomEnd(toDateInputValue(nextRange.end));
-    }
-  }, [periodMode, calendarDate]);
-
-  useEffect(() => {
-    if (periodMode !== "custom") return;
-    const start = startOfDay(fromDateInputValue(customStart, range.start));
-    const end = endOfDay(fromDateInputValue(customEnd, range.end));
-
-    if (start <= end) {
-      setRange({ start, end });
-      setCalendarDate(start);
-    }
-  }, [customStart, customEnd, periodMode]);
-
-  const { data: professionals = [] } = useQuery<Professional[]>({
-    queryKey: ["agenda-professionals", establishmentId, isEmployee, professionalId],
+  const { data, isLoading, error } = useQuery({
+    queryKey: ["agenda", establishmentId, range.start.toISOString(), range.end.toISOString()],
     enabled: !!establishmentId,
     queryFn: async () => {
-      let query = supabase
-        .from("professionals")
-        .select("id, name")
-        .eq("establishment_id", establishmentId)
-        .eq("active", true)
-        .order("name");
-
-      if (isEmployee && professionalId) {
-        query = query.eq("id", professionalId);
-      }
-
-      const { data, error } = await query;
-      if (error) throw error;
-      return (data ?? []) as Professional[];
-    },
-  });
-
-  useEffect(() => {
-    if (isEmployee || selectedProfessionalId === ALL_PROFESSIONALS || professionals.length === 0) return;
-    if (!professionals.some(professional => professional.id === selectedProfessionalId)) {
-      setSelectedProfessionalId(ALL_PROFESSIONALS);
-    }
-  }, [isEmployee, professionals, selectedProfessionalId]);
-
-  const { data, isLoading, isFetching, error } = useQuery({
-    queryKey: ["agenda", establishmentId, range.start.toISOString(), range.end.toISOString(), effectiveProfessionalId],
-    enabled: !!establishmentId && (!isEmployee || !!professionalId),
-    queryFn: async () => {
-      const appointmentFields = "id, establishment_id, appointment_date, status, notes, client_id, service_id, professional_id";
-      const appointmentRangeQuery = () => supabase
-        .from("appointments")
-        .select(appointmentFields)
-        .eq("establishment_id", establishmentId)
-        .gte("appointment_date", range.start.toISOString())
-        .lte("appointment_date", range.end.toISOString())
-        .order("appointment_date", { ascending: true });
-
-      const fetchAppointmentsByProfessional = async () => {
-        const primaryRes = await appointmentRangeQuery().eq("professional_id", effectiveProfessionalId);
-        if (primaryRes.error) return primaryRes;
-
-        const linkedIdsRes = await supabase
-          .from("appointment_professionals")
-          .select("appointment_id")
+      const [apptRes, servicesRes, profRes, clientsRes] = await Promise.all([
+        supabase.from("appointments").select("id, establishment_id, appointment_date, status, notes, client_id, service_id, professional_id")
           .eq("establishment_id", establishmentId)
-          .eq("professional_id", effectiveProfessionalId);
-
-        if (linkedIdsRes.error) return linkedIdsRes;
-
-        const linkedAppointmentIds = [
-          ...new Set((linkedIdsRes.data ?? []).map((row: any) => row.appointment_id).filter(Boolean)),
-        ];
-        const linkedAppointments: any[] = [];
-
-        for (let i = 0; i < linkedAppointmentIds.length; i += APPOINTMENT_ID_BATCH_SIZE) {
-          const ids = linkedAppointmentIds.slice(i, i + APPOINTMENT_ID_BATCH_SIZE);
-          const linkedRes = await supabase
-            .from("appointments")
-            .select(appointmentFields)
-            .eq("establishment_id", establishmentId)
-            .in("id", ids)
-            .gte("appointment_date", range.start.toISOString())
-            .lte("appointment_date", range.end.toISOString())
-            .order("appointment_date", { ascending: true });
-
-          if (linkedRes.error) return linkedRes;
-          linkedAppointments.push(...(linkedRes.data ?? []));
-        }
-
-        const byId = new Map<string, any>();
-        [...(primaryRes.data ?? []), ...linkedAppointments].forEach((appointment: any) => {
-          byId.set(appointment.id, appointment);
-        });
-
-        return {
-          data: [...byId.values()].sort(
-            (a, b) => new Date(a.appointment_date).getTime() - new Date(b.appointment_date).getTime()
-          ),
-          error: null,
-        };
-      };
-
-      const appointmentsPromise = effectiveProfessionalId
-        ? fetchAppointmentsByProfessional()
-        : appointmentRangeQuery();
-
-      let professionalsQuery = supabase
-        .from("professionals")
-        .select("id, name")
-        .eq("establishment_id", establishmentId)
-        .eq("active", true)
-        .order("name");
-
-      if (isEmployee && professionalId) {
-        professionalsQuery = professionalsQuery.eq("id", professionalId);
-      }
-
-      const [apptRes, servicesRes, profRes] = await Promise.all([
-        appointmentsPromise,
+          .gte("appointment_date", range.start.toISOString())
+          .lte("appointment_date", range.end.toISOString())
+          .order("appointment_date", { ascending: true }),
         supabase.from("services").select("id, name, duration_minutes").eq("establishment_id", establishmentId).eq("active", true),
-        professionalsQuery,
+        supabase.from("professionals").select("id, name").eq("establishment_id", establishmentId).eq("active", true),
+        supabase.from("clients").select("id, name").eq("establishment_id", establishmentId).order("name"),
       ]);
-
       if (apptRes.error) throw apptRes.error;
-      if (servicesRes.error) throw servicesRes.error;
-      if (profRes.error) throw profRes.error;
-
       const appts = apptRes.data ?? [];
       const services = servicesRes.data ?? [];
-      const activeProfessionals = (profRes.data ?? []) as Professional[];
-      const clientIds = [...new Set(appts.map((appt: any) => appt.client_id).filter(Boolean))];
-      const clientsRes = clientIds.length
-        ? await supabase.from("clients").select("id, name").in("id", clientIds)
-        : { data: [], error: null };
-
-      if (clientsRes.error) throw clientsRes.error;
-
+      const professionals = profRes.data ?? [];
+      const clients = clientsRes.data ?? [];
       const serviceMap = new Map(services.map((s: any) => [s.id, s]));
-      const profMap = new Map(activeProfessionals.map((p: any) => [p.id, p.name]));
-      const clientMap = new Map(((clientsRes.data ?? []) as any[]).map((c: any) => [c.id, c.name]));
-      return { appts, serviceMap, profMap, clientMap, services, professionals: activeProfessionals };
+      const profMap = new Map(professionals.map((p: any) => [p.id, p.name]));
+      const clientMap = new Map(clients.map((c: any) => [c.id, c.name]));
+      return { appts, serviceMap, profMap, clientMap, services, professionals, clients };
     },
   });
 
@@ -255,10 +71,6 @@ export default function AgendaContent() {
       return { id: a.id, title: `${client} · ${sname}`, start, end, status: a.status, raw: a };
     });
   }, [data]);
-
-  const selectedProfessionalName = effectiveProfessionalId
-    ? professionals.find(professional => professional.id === effectiveProfessionalId)?.name ?? "Profissional"
-    : "Todos os profissionais";
 
   const slug = (profile as any)?.slug as string | undefined;
   const publicLink = establishmentId ? `${window.location.origin}/${slug ?? `agendar/${establishmentId}`}` : "";
@@ -278,44 +90,6 @@ export default function AgendaContent() {
     const { error } = await supabase.from("appointments").update({ status }).eq("id", id);
     if (error) { toast({ title: "Erro", description: error.message, variant: "destructive" }); return; }
     refresh();
-  };
-
-  const handlePeriodChange = (value: string) => {
-    const nextPeriod = value as PeriodMode;
-    setPeriodMode(nextPeriod);
-
-    if (nextPeriod === "day") {
-      setCalendarView("day");
-    } else if (nextPeriod === "month") {
-      setCalendarView("month");
-    } else if (nextPeriod === "custom") {
-      setCalendarView("agenda");
-    } else {
-      setCalendarView("week");
-    }
-  };
-
-  const handleCalendarViewChange = (view: View) => {
-    setCalendarView(view);
-    if (view === "day" || view === "week" || view === "month") {
-      setPeriodMode(view);
-    }
-  };
-
-  const handleCalendarRangeChange = (nextRange: { start: Date; end: Date }) => {
-    if (periodMode === "custom") return;
-    setRange(nextRange);
-    setCustomStart(toDateInputValue(nextRange.start));
-    setCustomEnd(toDateInputValue(nextRange.end));
-  };
-
-  const navigateCustomRange = (direction: "previous" | "next") => {
-    const days = Math.max(1, Math.round((range.end.getTime() - range.start.getTime()) / 86_400_000) + 1);
-    const move = direction === "next" ? addDays : subDays;
-    const nextStart = move(range.start, days);
-    const nextEnd = move(range.end, days);
-    setCustomStart(toDateInputValue(nextStart));
-    setCustomEnd(toDateInputValue(nextEnd));
   };
 
   if (!establishmentId) return <div className="rounded-md border p-6 bg-card text-sm text-muted-foreground">Carregando perfil...</div>;
@@ -338,65 +112,6 @@ export default function AgendaContent() {
         </div>
       </div>
 
-      <div className="rounded-md border bg-card p-4 space-y-4">
-        <div className="grid gap-4 md:grid-cols-[minmax(220px,1fr)_220px_minmax(260px,1fr)]">
-          <div className="space-y-2">
-            <Label>Profissional</Label>
-            <Select
-              value={isEmployee ? professionalId ?? "" : selectedProfessionalId}
-              onValueChange={setSelectedProfessionalId}
-              disabled={isEmployee}
-            >
-              <SelectTrigger>
-                <SelectValue placeholder="Selecione um profissional" />
-              </SelectTrigger>
-              <SelectContent>
-                {!isEmployee && <SelectItem value={ALL_PROFESSIONALS}>Todos os profissionais</SelectItem>}
-                {professionals.map(professional => (
-                  <SelectItem key={professional.id} value={professional.id}>{professional.name}</SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-
-          <div className="space-y-2">
-            <Label>Período</Label>
-            <Select value={periodMode} onValueChange={handlePeriodChange}>
-              <SelectTrigger>
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="day">Dia</SelectItem>
-                <SelectItem value="week">Semana</SelectItem>
-                <SelectItem value="month">Mês</SelectItem>
-                <SelectItem value="custom">Intervalo personalizado</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-
-          <div className="space-y-2">
-            <Label>{periodMode === "custom" ? "Intervalo" : "Período atual"}</Label>
-            {periodMode === "custom" ? (
-              <div className="flex flex-wrap items-center gap-2">
-                <Input className="w-[145px]" type="date" value={customStart} onChange={(event) => setCustomStart(event.target.value)} />
-                <span className="text-sm text-muted-foreground">até</span>
-                <Input className="w-[145px]" type="date" value={customEnd} onChange={(event) => setCustomEnd(event.target.value)} />
-                <Button type="button" variant="outline" size="sm" onClick={() => navigateCustomRange("previous")}>Anterior</Button>
-                <Button type="button" variant="outline" size="sm" onClick={() => navigateCustomRange("next")}>Próximo</Button>
-              </div>
-            ) : (
-              <div className="flex min-h-10 items-center rounded-md border bg-muted/40 px-3 text-sm">
-                {formatRangeLabel(range.start, range.end)}
-              </div>
-            )}
-          </div>
-        </div>
-        <div className="flex flex-wrap items-center justify-between gap-2 text-sm text-muted-foreground">
-          <span>Exibindo {events.length} agendamento(s) para <strong className="text-foreground">{selectedProfessionalName}</strong>.</span>
-          {isFetching && <span>Atualizando agenda...</span>}
-        </div>
-      </div>
-
       {isLoading ? (
         <div className="rounded-md border p-6 bg-card text-sm text-muted-foreground">Carregando agendamentos...</div>
       ) : error ? (
@@ -404,14 +119,9 @@ export default function AgendaContent() {
       ) : viewMode === "calendar" ? (
         <AgendaCalendar
           events={events}
-          view={calendarView}
-          date={calendarDate}
-          agendaLength={Math.max(1, Math.round((range.end.getTime() - range.start.getTime()) / 86_400_000) + 1)}
-          onViewChange={handleCalendarViewChange}
-          onNavigate={setCalendarDate}
           onSelectSlot={handleSlot}
           onSelectEvent={handleEventClick}
-          onRangeChange={handleCalendarRangeChange}
+          onRangeChange={(r) => setRange(r)}
         />
       ) : (
         <div className="rounded-md border overflow-x-auto">
